added delete functionality to updating books.
When a book is selected from the home page, the book update page is rendered.  When the user submits the delete book button, they are prompted if they would like to do so.  If they select yes, the current book is deleted from the library database and they are redirected to the books page.